move link validation from form to model
